split code update, and delete
