+++ home.html
@@ -26,7 +26,7 @@
                     <div class="collapse navbar-collapse" id="navbarCollapse">
                         <ul class="navbar-nav me-auto mb-2 mb-md-0">
                             <li class="nav-item">
-                                <a class="nav-link active" aria-current="page" href="#">Home</a>
+                                <a class="nav-link active" aria-current="page" href="home.html">Home</a>
                             </li>
                             <li class="nav-item">
                                 <a class="nav-link" href="profile.html">Profile</a>
@@ -49,17 +49,18 @@
                 <h1>Create new post</h1>
                 <form class="form-post postForm">
                     <div class="mb-4">
-                        <label for="message-text" class="col-form-label">Post Title: <input name="title"
-                                onkeyup="success()" id="title" class="title form-control" /></label>
+                        <label for="message-text" class="col-form-label">Post Title: <input name="title" id="title"
+                                class="title form-control" /></label>
 
                     </div>
                     <div class="mb-4">
                         <label for="message-text" class="col-form-label">What would you like to say?</label>
-                        <textarea class="form-control textCreate" onkeyup="success()" id="message-text"></textarea>
+                        <textarea class="form-control textCreate" id="message-text"></textarea>
                     </div>
                     <div class="d-flex justify-content-end">
-                        <div class="addContent"></div>
-                        <button class="w-25 btn btn-sm btn-primary postBtn" type="submit">Create Post</button>
+                        <div class="addContent"><small>Title and content requireds</small></div>
+                        <button disabled="true" class="w-25 btn btn-sm btn-primary postBtn" type="submit">Create
+                            Post</button>
                     </div>
                 </form>
             </div>
